third_party/blink/web_tests/platform/generic/fast/dom/HTMLTemplateElement/cycles-expected.txt

Test that cycles are not allowed in template content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS template.content.appendChild(template) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS template.content.appendChild(outerDiv) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS innerDiv.appendChild(template) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS innerDiv.appendChild(outerDiv) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS innerTemplate.appendChild(outerDiv) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
